scripts/requirements-logo.txt
# Dependencies for logo generation
Pillow>=9.0.0
